requirements.txt
pyTelegramBotAPI==4.12.0
numpy==1.23.5
PyYAML==5.4.1
google-cloud-storage==2.7.0
opencv-python-headless==4.8.0.74
onnxruntime==1.15.1
insightface==0.7.3
faiss-cpu== 1.7.4
SQLAlchemy==2.0.17
sqlalchemy-pytds==1.0.0
google-cloud-core==2.3.2
pg8000==1.29.8
cloud-sql-python-connector==1.2.4


